# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документация для API",
        "contact": {
            "email": "alekscygankov20@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/password-saver-api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получение информации о текущем пользователе",
                "operationId": "userInfoPasswordSaverApi",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID пользователя",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "Имя пользователя",
                                                    "type": "string",
                                                    "example": "TestName"
                                                },
                                                "email": {
                                                    "description": "Email пользователя",
                                                    "type": "string",
                                                    "example": "test@gmail.com"
                                                },
                                                "active": {
                                                    "description": "Активен ли аккаунт пользователя",
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "token": {
                                                    "description": "Текущий Bearer токен",
                                                    "type": "string",
                                                    "example": "5|EfiojPaSzOLru1uqi7A9YlA3xggl6tCWkoKHWrVkf3d74066"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Пользователь не аутентифицирован",
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/password-saver-api/user/setPinCode": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Установить PIN-код для пользователя",
                "operationId": "setUserPinCodePasswordSaverApi",
                "requestBody": {
                    "description": "Данные для установки PIN-кода",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pin_code"
                                ],
                                "properties": {
                                    "pin_code": {
                                        "description": "4-значный PIN-код пользователя",
                                        "type": "integer",
                                        "example": 1234
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PIN-код успешно установлен."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Пользователь не аутентифицирован",
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Текст ошибки валидации"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/password-saver-api/user/updatePinCode": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Обновить PIN-код пользователя",
                "operationId": "updateUserPinCodePasswordSaverApi",
                "requestBody": {
                    "description": "Данные для обновления PIN-кода",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pin_code",
                                    "old_pin_code"
                                ],
                                "properties": {
                                    "pin_code": {
                                        "description": "Новый 4-значный PIN-код пользователя",
                                        "type": "integer",
                                        "example": 4321
                                    },
                                    "old_pin_code": {
                                        "description": "Текущий 4-значный PIN-код пользователя для подтверждения",
                                        "type": "integer",
                                        "example": 1234
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PIN-код успешно обнавлён."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Пользователь не аутентифицирован",
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Текст ошибки валидации"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}